tv_input: fix 720i width rsolution happen blurred screen
need the commit of:
  https://10.10.10.29/c/android/rk/hardware/libgralloc/bifrost/+/158188
  calc_allocation_size(): Adjust behaviors of "adjusting byte_strde of buffers with format_used_by_rk_video

+++ common/TvInput_Buffer_Manager.h
@@ -94,7 +94,7 @@ class EXPORTED TvInputBufferManager {
   virtual int Allocate(size_t width,
                        size_t height,
                        uint32_t format,
-                       uint32_t usage,
+                       uint64_t usage,
                        BufferType type,
                        buffer_handle_t* out_buffer,
                        uint32_t* out_stride) = 0;
